pyproject.toml
[project]
name = "cocacola-text-2-sql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.10.1",
    "aiohappyeyeballs>=2.6.1",
    "aiohttp>=3.13.0",
    "aiosignal>=1.4.0",
    "altair>=5.5.0",
    "annotated-types>=0.7.0",
    "anyio>=4.11.0",
    "arro3-core>=0.6.3",
    "asttokens>=3.0.0",
    "attrs>=25.4.0",
    "blinker>=1.9.0",
    "cachetools>=6.2.0",
    "certifi>=2025.10.5",
    "cffi>=2.0.0",
    "charset-normalizer>=3.4.3",
    "choreographer>=1.1.2",
    "click>=8.3.0",
    "cohere>=5.18.0",
    "colorama>=0.4.6",
    "comm>=0.2.3",
    "commentjson>=0.9.0",
    "contourpy>=1.3.3",
    "cryptography>=46.0.2",
    "cycler>=0.12.1",
    "dataclasses-json>=0.6.7",
    "debugpy>=1.8.17",
    "decorator>=5.2.1",
    "deprecation>=2.1.0",
    "diffusers>=0.9.0",
    "diskcache>=5.6.3",
    "distro>=1.9.0",
    "executing>=2.2.1",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.119.0",
    "fastavro>=1.12.1",
    "fastjsonschema>=2.21.2",
    "filelock>=3.20.0",
    "filetype>=1.2.0",
    "fonttools>=4.60.1",
    "frozenlist>=1.8.0",
    "fsspec>=2025.9.0",
    "ftfy>=6.3.1",
    "geopandas>=1.1.1",
    "gitdb>=4.0.12",
    "gitpython>=3.1.45",
    "google-ai-generativelanguage>=0.7.0",
    "google-api-core>=2.26.0",
    "google-auth>=2.41.1",
    "googleapis-common-protos>=1.70.0",
    "gotrue>=2.12.4",
    "greenlet>=3.2.4",
    "grpcio>=1.75.1",
    "grpcio-status>=1.75.1",
    "h11>=0.16.0",
    "h2>=4.3.0",
    "hpack>=4.1.0",
    "httpcore>=1.0.9",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.0",
    "huggingface-hub>=0.35.3",
    "hyperframe>=6.1.0",
    "idna>=3.10",
    "importlib-metadata>=8.7.0",
    "ipykernel>=6.30.1",
    "ipython>=9.6.0",
    "ipython-pygments-lexers>=1.1.1",
    "jedi>=0.19.2",
    "jinja2>=3.1.6",
    "jiter>=0.11.0",
    "jsonpatch>=1.33",
    "jsonpointer>=3.0.0",
    "jsonref>=1.1.0",
    "jsonschema>=4.25.1",
    "jsonschema-specifications>=2025.9.1",
    "jupyter-client>=8.6.3",
    "jupyter-core>=5.8.1",
    "kaleido>=1.1.0",
    "kiwisolver>=1.4.9",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-core>=0.3.79",
    "langchain-google-genai>=2.1.12",
    "langchain-openai>=0.3.35",
    "langchain-text-splitters>=0.3.11",
    "langgraph>=0.6.10",
    "langgraph-checkpoint>=2.1.2",
    "langgraph-prebuilt>=0.6.4",
    "langgraph-sdk>=0.2.9",
    "langsmith>=0.4.34",
    "lark-parser>=0.7.8",
    "lida>=0.0.14",
    "llmx>=0.0.21a0",
    "logistro>=1.1.0",
    "markdown-it-py>=4.0.0",
    "markupsafe>=3.0.3",
    "marshmallow>=3.26.1",
    "matplotlib>=3.10.7",
    "matplotlib-inline>=0.1.7",
    "matplotlib-venn>=1.1.2",
    "mdurl>=0.1.2",
    "mizani>=0.14.2",
    "mpmath>=1.3.0",
    "multidict>=6.7.0",
    "mypy-extensions>=1.1.0",
    "mysql-connector>=2.2.9",
    "mysql-connector-python>=9.4.0",
    "narwhals>=2.7.0",
    "nbformat>=5.10.4",
    "nest-asyncio>=1.6.0",
    "networkx>=3.5",
    "numpy>=2.3.3",
    "openai>=2.3.0",
    "opencv-python>=4.11.0.86",
    "orjson>=3.11.3",
    "ormsgpack>=1.11.0",
    "packaging>=25.0",
    "pandas>=2.3.3",
    "parso>=0.8.5",
    "patsy>=1.0.1",
    "peacasso>=0.0.19a0",
    "pillow>=11.3.0",
    "platformdirs>=4.5.0",
    "plotly>=6.3.1",
    "plotnine>=0.15.0",
    "postgrest>=2.22.0",
    "prompt-toolkit>=3.0.52",
    "propcache>=0.4.1",
    "proto-plus>=1.26.1",
    "protobuf>=6.32.1",
    "psutil>=7.1.0",
    "pure-eval>=0.2.3",
    "pyarrow>=21.0.0",
    "pyasn1>=0.6.1",
    "pyasn1-modules>=0.4.2",
    "pycparser>=2.23",
    "pydantic>=2.12.0",
    "pydantic-core>=2.41.1",
    "pydantic-settings>=2.11.0",
    "pydeck>=0.9.1",
    "pygments>=2.19.2",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.2",
    "pyodbc>=5.2.0",
    "pyogrio>=0.11.1",
    "pyparsing>=3.2.5",
    "pyproj>=3.7.2",
    "python-dateutil>=2.9.0.post0",
    "python-decouple>=3.8",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "pytz>=2025.2",
    "pywin32>=311",
    "pyyaml>=6.0.3",
    "pyzmq>=27.1.0",
    "rapidfuzz>=3.14.1",
    "realtime>=2.22.0",
    "referencing>=0.36.2",
    "regex>=2025.9.18",
    "requests>=2.32.5",
    "requests-toolbelt>=1.0.0",
    "rich>=14.2.0",
    "rpds-py>=0.27.1",
    "rsa>=4.9.1",
    "safetensors>=0.6.2",
    "scipy>=1.16.2",
    "seaborn>=0.13.2",
    "setuptools>=80.9.0",
    "shapely>=2.1.2",
    "shellingham>=1.5.4",
    "simplejson>=3.20.2",
    "six>=1.17.0",
    "smmap>=5.0.2",
    "sniffio>=1.3.1",
    "sqlalchemy>=2.0.44",
    "stack-data>=0.6.3",
    "starlette>=0.48.0",
    "statsmodels>=0.14.5",
    "storage3>=2.22.0",
    "streamlit>=1.50.0",
    "streamlit-aggrid>=1.1.9",
    "streamlit-chat>=0.1.1",
    "streamlit-option-menu>=0.4.0",
    "strenum>=0.4.15",
    "supabase>=2.22.0",
    "supafunc>=0.10.2",
    "sympy>=1.14.0",
    "tabulate>=0.9.0",
    "tenacity>=9.1.2",
    "thefuzz>=0.22.1",
    "tiktoken>=0.12.0",
    "tokenizers>=0.22.1",
    "toml>=0.10.2",
    "torch>=2.8.0",
    "tornado>=6.5.2",
    "tqdm>=4.67.1",
    "traitlets>=5.14.3",
    "transformers>=4.57.0",
    "typer>=0.19.2",
    "types-requests>=2.32.4.20250913",
    "typing-extensions>=4.15.0",
    "typing-inspect>=0.9.0",
    "typing-inspection>=0.4.2",
    "tzdata>=2025.2",
    "urllib3>=2.5.0",
    "uvicorn>=0.37.0",
    "vegafusion>=2.0.3",
    "watchdog>=6.0.0",
    "wcwidth>=0.2.14",
    "websockets>=15.0.1",
    "wordcloud>=1.9.4",
    "xxhash>=3.6.0",
    "yarl>=1.22.0",
    "zipp>=3.23.0",
    "zstandard>=0.25.0",
]
